Move attribute completion tests
