+++ crates/ide_completion/src/completions/attribute/derive.rs
@@ -77,93 +77,3 @@ const DEFAULT_DERIVE_COMPLETIONS: &[DeriveDependencies] = &[
     DeriveDependencies { label: "Ord", dependencies: &["PartialOrd", "Eq", "PartialEq"] },
     DeriveDependencies { label: "PartialOrd", dependencies: &["PartialEq"] },
 ];
-
-#[cfg(test)]
-mod tests {
-    use expect_test::{expect, Expect};
-
-    use crate::tests::completion_list;
-
-    fn check(ra_fixture: &str, expect: Expect) {
-        let builtin_derives = r#"
-#[rustc_builtin_macro]
-pub macro Clone {}
-#[rustc_builtin_macro]
-pub macro Copy {}
-#[rustc_builtin_macro]
-pub macro Default {}
-#[rustc_builtin_macro]
-pub macro Debug {}
-#[rustc_builtin_macro]
-pub macro Hash {}
-#[rustc_builtin_macro]
-pub macro PartialEq {}
-#[rustc_builtin_macro]
-pub macro Eq {}
-#[rustc_builtin_macro]
-pub macro PartialOrd {}
-#[rustc_builtin_macro]
-pub macro Ord {}
-
-"#;
-        let actual = completion_list(&format!("{} {}", builtin_derives, ra_fixture));
-        expect.assert_eq(&actual);
-    }
-
-    #[test]
-    fn no_completion_for_incorrect_derive() {
-        check(r#"#[derive{$0)] struct Test;"#, expect![[]])
-    }
-
-    #[test]
-    fn empty_derive() {
-        check(
-            r#"#[derive($0)] struct Test;"#,
-            expect![[r#"
-            at PartialEq
-            at Default
-            at PartialEq, Eq
-            at PartialEq, Eq, PartialOrd, Ord
-            at Clone, Copy
-            at Debug
-            at Clone
-            at Hash
-            at PartialEq, PartialOrd
-        "#]],
-        );
-    }
-
-    #[test]
-    fn derive_with_input() {
-        check(
-            r#"#[derive(serde::Serialize, PartialEq, $0)] struct Test;"#,
-            expect![[r#"
-                at Default
-                at Eq
-                at Eq, PartialOrd, Ord
-                at Clone, Copy
-                at Debug
-                at Clone
-                at Hash
-                at PartialOrd
-            "#]],
-        )
-    }
-
-    #[test]
-    fn derive_with_input2() {
-        check(
-            r#"#[derive($0 serde::Serialize, PartialEq)] struct Test;"#,
-            expect![[r#"
-                at Default
-                at Eq
-                at Eq, PartialOrd, Ord
-                at Clone, Copy
-                at Debug
-                at Clone
-                at Hash
-                at PartialOrd
-            "#]],
-        )
-    }
-}
