+++ crates/ide_completion/src/completions/attribute/lint.rs
@@ -30,44 +30,3 @@ pub(super) fn complete_lint(
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::tests::check_edit;
-
-    #[test]
-    fn check_empty() {
-        check_edit(
-            "deprecated",
-            r#"#[allow($0)] struct Test;"#,
-            r#"#[allow(deprecated)] struct Test;"#,
-        )
-    }
-
-    #[test]
-    fn check_with_existing() {
-        check_edit(
-            "deprecated",
-            r#"#[allow(keyword_idents, $0)] struct Test;"#,
-            r#"#[allow(keyword_idents, deprecated)] struct Test;"#,
-        )
-    }
-
-    #[test]
-    fn check_qualified() {
-        check_edit(
-            "deprecated",
-            r#"#[allow(keyword_idents, $0)] struct Test;"#,
-            r#"#[allow(keyword_idents, deprecated)] struct Test;"#,
-        )
-    }
-
-    #[test]
-    fn check_feature() {
-        check_edit(
-            "box_syntax",
-            r#"#[feature(box_$0)] struct Test;"#,
-            r#"#[feature(box_syntax)] struct Test;"#,
-        )
-    }
-}
