+++ crates/ide_completion/src/completions/attribute/repr.rs
@@ -68,132 +68,3 @@ const REPR_COMPLETIONS: &[ReprCompletion] = &[
     attr("i28",    &["transparent", "u8", "u16", "u32", "u64", "u128", "usize", "i8", "i16", "i32", "i64", "isize"]),
     attr("isize",  &["transparent", "u8", "u16", "u32", "u64", "u128", "usize", "i8", "i16", "i32", "i64", "i128"]),
 ];
-
-#[cfg(test)]
-mod tests {
-    use expect_test::{expect, Expect};
-
-    use crate::tests::completion_list;
-
-    fn check(ra_fixture: &str, expect: Expect) {
-        let actual = completion_list(ra_fixture);
-        expect.assert_eq(&actual);
-    }
-
-    #[test]
-    fn no_completion_for_incorrect_repr() {
-        check(r#"#[repr{$0)] struct Test;"#, expect![[]])
-    }
-
-    #[test]
-    fn empty() {
-        check(
-            r#"#[repr($0)] struct Test;"#,
-            expect![[r#"
-                at align($0)
-                at packed
-                at transparent
-                at C
-                at u8
-                at u16
-                at u32
-                at u64
-                at u128
-                at usize
-                at i8
-                at i16
-                at i32
-                at i64
-                at i28
-                at isize
-            "#]],
-        );
-    }
-
-    #[test]
-    fn transparent() {
-        check(r#"#[repr(transparent, $0)] struct Test;"#, expect![[r#""#]]);
-    }
-
-    #[test]
-    fn align() {
-        check(
-            r#"#[repr(align(1), $0)] struct Test;"#,
-            expect![[r#"
-                at align($0)
-                at transparent
-                at C
-                at u8
-                at u16
-                at u32
-                at u64
-                at u128
-                at usize
-                at i8
-                at i16
-                at i32
-                at i64
-                at i28
-                at isize
-            "#]],
-        );
-    }
-
-    #[test]
-    fn packed() {
-        check(
-            r#"#[repr(packed, $0)] struct Test;"#,
-            expect![[r#"
-                at transparent
-                at C
-                at u8
-                at u16
-                at u32
-                at u64
-                at u128
-                at usize
-                at i8
-                at i16
-                at i32
-                at i64
-                at i28
-                at isize
-            "#]],
-        );
-    }
-
-    #[test]
-    fn c() {
-        check(
-            r#"#[repr(C, $0)] struct Test;"#,
-            expect![[r#"
-                at align($0)
-                at packed
-                at u8
-                at u16
-                at u32
-                at u64
-                at u128
-                at usize
-                at i8
-                at i16
-                at i32
-                at i64
-                at i28
-                at isize
-            "#]],
-        );
-    }
-
-    #[test]
-    fn prim() {
-        check(
-            r#"#[repr(usize, $0)] struct Test;"#,
-            expect![[r#"
-                at align($0)
-                at packed
-                at C
-            "#]],
-        );
-    }
-}
